requirements.txt
accelerate==1.5.1
aicsimageio==4.14.0
aiobotocore==2.5.4
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==0.7.16
annotated-types==0.7.0
app-model==0.3.1
appdirs==1.4.4
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
blessed==1.20.0
botocore==1.31.17
build==1.2.2.post1
cachey==0.2.1
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dask==2024.8.0
debugpy==1.8.9
decorator==5.1.1
diffusers==0.32.2
distributed==2024.8.0
docstring_parser==0.16
docutils==0.21.2
elementpath==4.6.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
facenet-pytorch==2.6.0
fasteners==0.19
fastmri==0.3.0
filelock==3.16.1
flexcache==0.3
flexparser==0.4
fonttools==4.54.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2023.6.0
gdown==5.2.0
gpustat==1.1.1
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
HeapDict==1.0.1
hsluv==5.0.4
huggingface-hub==0.29.3
imagecodecs==2024.9.22
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
in-n-out==0.2.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
llvmlite==0.43.0
locket==1.0.0
lxml==4.9.4
magicgui==0.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
napari==0.5.5
napari-console==0.1.2
napari-plugin-engine==0.2.0
napari-svg==0.2.0
nest-asyncio==1.6.0
networkx==3.2.1
npe2==0.7.7
numba==0.60.0
numcodecs==0.12.1
numpy==1.24.0
numpydoc==1.8.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
ome-types==0.5.3
ome-zarr==0.9.0
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.2.0
Pint==0.24.4
platformdirs==4.3.6
playwright==1.49.1
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.1
psygnal==0.11.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyconify==0.1.6
pydantic==2.10.3
pydantic-compat==0.1.2
pydantic_core==2.27.1
pyee==12.0.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
pyproject_hooks==1.2.0
pyspng==0.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-fid==0.3.0
pytorch-lightning==2.5.0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.2
referencing==0.35.1
regex==2024.11.6
resource-backed-dask-array==0.1.0
rich==13.9.4
rpds-py==0.22.3
runstats==2.0.0
s3fs==2023.6.0
safetensors==0.5.3
scikit-image==0.24.0
selenium==4.27.1
shellingham==1.5.4
sigpy==0.1.27
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
superqt==0.6.7
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
threadpoolctl==3.6.0
tifffile==2023.2.28
tomli==2.2.1
tomli_w==1.1.0
toolz==1.0.0
torch==2.2.2
torchmetrics==1.6.0
torchvision==0.17.2
tornado==6.4.2
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
triton==2.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vispy==0.14.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
xarray==2024.7.0
xmlschema==3.4.3
xsdata==24.3.1
yarl==1.18.3
zarr==2.15.0
zict==3.0.0
zipp==3.20.2
